AWS_noisy_label_detection/requirements.txt
fastapi
uvicorn
pillow
python-multipart
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4